Plans: fix the raw price
Using the raw price coming from the plans endpoint instead of the site
plans endpoint leads to obscure bugs if discounts are applied globally.

It's possible that a discount applies to the current site but does not
apply to any other site, so plans would return, let's say 100, and
sites/plans would return 150.