Display the crossed prices on signup as well
This change introduces a new plan selector - getDiscountedRawPrice. The
naming selection for it was difficult because of the very inconsistent
past naming selections.

We have:
- getSitePlanRawPrice
- getPlanDiscountedRawPrice // should be getSitePlanDiscountedRawPrice
- getPlanRawPrice
- getDiscountedRawPrice // should be getPlanDiscountedRawPrice

I didn't want to rename functions in this already large PR, resolving
the inconsistency in naming should be planned separately. I would
suggest that selectors for the state.plans are prefixed with getPlan and
selectors for state.sites.plans are prefixed with getSitePlan.